:white_check_mark: Adds Tests for Size Reactivity
