+++ packages/reactivity/src/reactive/collectionMethods.ts
@@ -1,4 +1,6 @@
+import { Effect } from '../Effect';
 import { Signal, untrack } from '../Signal';
+import { nextTick } from '../nextTick';
 import { MapTypes, hasOwn, isObject } from '../utils';
 import { proxyMap } from './proxyMap';
 import { reactive, toRaw, wrap } from './reactive';
@@ -29,6 +31,8 @@ const collectionTraps: ProxyHandler<MapTypes> = {
   },
 };
 
+const $SIZE = Symbol('$SIZE');
+
 const collectionMethods: Partial<
   Map<unknown, unknown> & WeakMap<object, unknown>
 > = {
@@ -51,19 +55,17 @@ const collectionMethods: Partial<
       target.set(rawKey as object, rawValue);
       return this;
     }
+    const sizeSignal = nodes.get($SIZE);
     if (nodes.has(rawKey)) {
-      const node = nodes.get(rawKey)!;
-      if (!untrack(() => Object.is(toRaw(node.get()), rawValue)))
-        node.set(isObject(rawValue) ? reactive(rawValue) : rawValue);
-      target.set(rawKey as object, rawValue);
-      return this;
+      const signal = nodes.get(rawKey)!;
+      if (!untrack(() => Object.is(toRaw(signal.get()), rawValue)))
+        signal.set(isObject(rawValue) ? reactive(rawValue) : rawValue);
+    } else {
+      const signal = wrap(rawValue);
+      nodes.set(rawKey, signal);
     }
-    const signal = wrap(rawValue);
-    nodes.set(rawKey, signal);
-    target.set(
-      rawKey as object,
-      untrack(() => signal.get()),
-    );
+    target.set(rawKey as object, rawValue);
+    if (sizeSignal) sizeSignal.set(Reflect.get(target, 'size', target));
     return this;
   },
   has(this: MapTypes, key: unknown) {
@@ -79,6 +81,7 @@ const collectionMethods: Partial<
     if (nodes.has(rawKey)) {
       nodes.get(rawKey)?.set(undefined);
       nodes.delete(rawKey);
+      nodes.get($SIZE)?.set(Reflect.get(target, 'size', target));
     }
     return target.delete(rawKey as object);
   },
@@ -93,6 +96,155 @@ const collectionMethods: Partial<
   },
   get size() {
     const target = toRaw(this);
-    return Reflect.get(target, 'size', target);
+    const nodes = reactiveNodes.get(target);
+    const innerSize = Reflect.get(target, 'size', target);
+    if (!nodes) return innerSize;
+    if (!nodes.has($SIZE)) {
+      const signal = new Signal(innerSize);
+      nodes.set($SIZE, signal);
+    }
+    const sizeSignal = nodes.get($SIZE)!;
+    sizeSignal.set(innerSize);
+    return sizeSignal.get();
   },
 };
+
+if (import.meta.vitest) {
+  describe('(Weak)Map', () => {
+    it('can handle Map', async () => {
+      const map = reactive(new Map());
+      map.set('foo', 42);
+      let value = 0;
+      new Effect(() => {
+        value = map.get('foo');
+      });
+      expect(value).toBe(42);
+      map.set('foo', 100);
+      expect(map.get('foo')).toBe(100);
+      await nextTick();
+      expect(value).toBe(100);
+    });
+    it('can handle nested objects inside Map with object keys', async () => {
+      const map = reactive(new Map());
+      const key = { foo: 'bar' };
+      map.set(key, {
+        bar: 42,
+      });
+      let value = 0;
+      new Effect(() => {
+        value = map.get(key).bar;
+      });
+      expect(value).toBe(42);
+      map.get(key).bar = 100;
+      expect(map.get(key).bar).toBe(100);
+      await nextTick();
+      expect(value).toBe(100);
+    });
+    it('can get keys before they are set', async () => {
+      const map = reactive(new Map());
+      const key = { foo: 'bar' };
+      let value = 0;
+      new Effect(() => {
+        value = map.get(key)?.bar;
+      });
+      expect(value).toBe(undefined);
+      map.set(key, {
+        bar: 42,
+      });
+      await nextTick();
+      expect(value).toBe(42);
+      map.get(key).bar = 100;
+      expect(map.get(key).bar).toBe(100);
+      await nextTick();
+      expect(value).toBe(100);
+    });
+    it('can safely use reactive objects as keys', () => {
+      const map = reactive(new Map());
+      const rawKey = { foo: 'bar' };
+      const key = reactive(rawKey);
+      map.set(key, 42);
+      expect(map.get(key)).toBe(42);
+      expect(map.get(rawKey)).toBe(42);
+    });
+    it('.has: can check if a key is present', () => {
+      const map = reactive(new Map());
+      const rawKey = { foo: 'bar' };
+      const key = reactive(rawKey);
+      expect(map.has(key)).toBe(false);
+      map.set(key, 42);
+      expect(map.has(rawKey)).toBe(true);
+    });
+    it('.delete: can delete a key', async () => {
+      const map = reactive(new Map());
+      let value = 0;
+      const key = 'foo';
+      new Effect(() => (value = map.get(key)));
+      expect(value).toBe(undefined);
+      map.set(key, 42);
+      await nextTick();
+      expect(map.get(key)).toBe(42);
+      expect(map.has(key)).toBe(true);
+      expect(value).toBe(42);
+      map.delete(key);
+      await nextTick();
+      expect(map.has(key)).toBe(false);
+      expect(map.get(key)).toBe(undefined);
+      expect(value).toBe(undefined);
+      map.set(key, 69);
+      await nextTick();
+      expect(map.get(key)).toBe(69);
+      expect(map.has(key)).toBe(true);
+      expect(value).toBe(69);
+    });
+    it('.clear: can clear all pairs', async () => {
+      const map = reactive(new Map());
+      const key1 = 'foo';
+      const key2 = 'bar';
+      let value = 0;
+      new Effect(() => (value = map.get(key1) + map.get(key2)));
+      map.set(key1, 42);
+      map.set(key2, 69);
+      await nextTick();
+      expect(map.has(key1)).toBe(true);
+      expect(map.has(key2)).toBe(true);
+      expect(value).toBe(111);
+      map.clear();
+      await nextTick();
+      expect(map.has(key1)).toBe(false);
+      expect(map.has(key2)).toBe(false);
+      expect(value).toBe(NaN);
+      map.set(key1, 420);
+      map.set(key2, 69);
+      await nextTick();
+      expect(map.has(key1)).toBe(true);
+      expect(map.has(key2)).toBe(true);
+      expect(value).toBe(489);
+    });
+    it('.size: can get the size', async () => {
+      const map = reactive(new Map());
+      const fn = vi.fn(() => (value = map.size));
+      let value: number = 42;
+      new Effect(fn);
+      expect(value).toBe(0);
+      expect(fn).toBeCalledTimes(1);
+      map.set('foo', 42);
+      await nextTick();
+      expect(value).toBe(1);
+      map.set('bar', 69);
+      await nextTick();
+      expect(map.size).toBe(2);
+      expect(value).toBe(2);
+      expect(fn).toBeCalledTimes(3);
+      map.set('foo', 69);
+      await nextTick();
+      expect(fn).toBeCalledTimes(4);
+      expect(value).toBe(2);
+      map.delete('foo');
+      await nextTick();
+      expect(value).toBe(1);
+      map.clear();
+      await nextTick();
+      expect(value).toBe(0);
+    });
+  });
+}
