+++ packages/reactivity/src/reactive/reactive.ts
@@ -269,117 +269,5 @@ if (import.meta.vitest) {
         expect(value).toBe(6);
       });
     });
-
-    describe('(Weak)Map', () => {
-      it('can handle Map', async () => {
-        const map = reactive(new Map());
-        map.set('foo', 42);
-        let value = 0;
-        new Effect(() => {
-          value = map.get('foo');
-        });
-        expect(value).toBe(42);
-        map.set('foo', 100);
-        expect(map.get('foo')).toBe(100);
-        await nextTick();
-        expect(value).toBe(100);
-      });
-      it('can handle nested objects inside Map with object keys', async () => {
-        const map = reactive(new Map());
-        const key = { foo: 'bar' };
-        map.set(key, {
-          bar: 42,
-        });
-        let value = 0;
-        new Effect(() => {
-          value = map.get(key).bar;
-        });
-        expect(value).toBe(42);
-        map.get(key).bar = 100;
-        expect(map.get(key).bar).toBe(100);
-        await nextTick();
-        expect(value).toBe(100);
-      });
-      it('can get keys before they are set', async () => {
-        const map = reactive(new Map());
-        const key = { foo: 'bar' };
-        let value = 0;
-        new Effect(() => {
-          value = map.get(key)?.bar;
-        });
-        expect(value).toBe(undefined);
-        map.set(key, {
-          bar: 42,
-        });
-        await nextTick();
-        expect(value).toBe(42);
-        map.get(key).bar = 100;
-        expect(map.get(key).bar).toBe(100);
-        await nextTick();
-        expect(value).toBe(100);
-      });
-      it('can safely use reactive objects as keys', () => {
-        const map = reactive(new Map());
-        const rawKey = { foo: 'bar' };
-        const key = reactive(rawKey);
-        map.set(key, 42);
-        expect(map.get(key)).toBe(42);
-        expect(map.get(rawKey)).toBe(42);
-      });
-      it('.has: can check if a key is present', () => {
-        const map = reactive(new Map());
-        const rawKey = { foo: 'bar' };
-        const key = reactive(rawKey);
-        expect(map.has(key)).toBe(false);
-        map.set(key, 42);
-        expect(map.has(rawKey)).toBe(true);
-      });
-      it('.delete: can delete a key', async () => {
-        const map = reactive(new Map());
-        let value = 0;
-        const key = 'foo';
-        new Effect(() => (value = map.get(key)));
-        expect(value).toBe(undefined);
-        map.set(key, 42);
-        await nextTick();
-        expect(map.get(key)).toBe(42);
-        expect(map.has(key)).toBe(true);
-        expect(value).toBe(42);
-        map.delete(key);
-        await nextTick();
-        expect(map.has(key)).toBe(false);
-        expect(map.get(key)).toBe(undefined);
-        expect(value).toBe(undefined);
-        map.set(key, 69);
-        await nextTick();
-        expect(map.get(key)).toBe(69);
-        expect(map.has(key)).toBe(true);
-        expect(value).toBe(69);
-      });
-      it('.clear: can clear all pairs', async () => {
-        const map = reactive(new Map());
-        const key1 = 'foo';
-        const key2 = 'bar';
-        let value = 0;
-        new Effect(() => (value = map.get(key1) + map.get(key2)));
-        map.set(key1, 42);
-        map.set(key2, 69);
-        await nextTick();
-        expect(map.has(key1)).toBe(true);
-        expect(map.has(key2)).toBe(true);
-        expect(value).toBe(111);
-        map.clear();
-        await nextTick();
-        expect(map.has(key1)).toBe(false);
-        expect(map.has(key2)).toBe(false);
-        expect(value).toBe(NaN);
-        map.set(key1, 420);
-        map.set(key2, 69);
-        await nextTick();
-        expect(map.has(key1)).toBe(true);
-        expect(map.has(key2)).toBe(true);
-        expect(value).toBe(489);
-      });
-    });
   });
 }
