+++ size.json
@@ -1,22 +1,22 @@
 {
   "reactivity": {
     "minified": {
-      "pretty": "3.14 kB",
-      "raw": 3144
+      "pretty": "3.32 kB",
+      "raw": 3319
     },
     "brotli": {
-      "pretty": "1.18 kB",
-      "raw": 1178
+      "pretty": "1.24 kB",
+      "raw": 1236
     }
   },
   "alpinejs": {
     "minified": {
-      "pretty": "34.4 kB",
-      "raw": 34372
+      "pretty": "34.5 kB",
+      "raw": 34546
     },
     "brotli": {
       "pretty": "11.6 kB",
-      "raw": 11551
+      "raw": 11601
     }
   }
 }
